Create README.md
This project delves into the effectiveness of gradient descent-based models compared to other machine learning models. To achieve this, we employ a host of models, including a Decision Tree, a Random Forest, a Gradient Boosting Classifier, a Stochastic Gradient Descent Classifier, and an XGBoost Classifier. These models are employed to predict a newly created categorical response variable that represents the likelihood of a student dropout in New York State school districts